+++ statik/statik.go
@@ -1,7 +1,7 @@
 package statik
 
 import (
-		"github.com/rakyll/statik/fs"
+	"github.com/rakyll/statik/fs"
 )
 
 func init() {